Reworked Damage System
Reworked how damages works against both the player and enemies

Player:
- Player no longer has health. Instead performance is measured by score.
- PointSystem script handles updating the current score.
- Score is gained by killing enemies. Each enemy has a different point value rewarded upon death.
- Points are also rewarded for landing shots on enemies. 100 for a critical hit, 25 for a body shot, and 10 for a limb shot.

Enemies:
- Enemies can now take True Damage. True Damage occurs for non-raycast sources of damage and does not have a multiplier affecting it.
- Enemies will catalogue sources of True Damage briefly upon taking damage. This is to make sure they are not affected by the same hurtbox every frame.

+++ Run and Gun/Assets/Scripts/Enemies/Infector.cs
@@ -19,7 +19,7 @@ public class Infector : BaseEnemy
     public float buffer = 10f;
 
     [Header("Bullet Settings")]
-    [SerializeField] private float attackDamage;
+    [SerializeField] private int attackDamage;
     [SerializeField] private float bulletSpeed;
 
     private void LateUpdate()
